feature: pulled the most recent dataset for Traits from GCS refactor: refactored Aspect subclassess to be intact with recent pull fix: fixed _title property of a Trait to actually replace the @...@ pattern and not hte whole name

+++ lib/widgets/compose_page/dialogs/select_trait_modifiers.dart
@@ -57,7 +57,7 @@ class _SelectTraitModifiersDialogState
 
   Widget _buildForm() {
     widget.trait.modifiers!.sort(
-      (TraitModifier a, TraitModifier b) => a.cost - b.cost,
+      (TraitModifier a, TraitModifier b) => a.cost.toInt() - b.cost.toInt(),
     );
 
     return Column(
